Added DevProjects to useful links and under AngularJS, Java, PHP, and Python
DevProjects is a community with free real-world projects in multiple coding languages and tech stacks. The website also includes solutions from other developers and discussion board for all users to participate in. Apart from the projects, solutions, and discussion board, users can also reach out to mentors on Codementor for paid help. However, the latter bit is not compulsory or forced.